# api/openapi.yml
openapi: '3.0.3'
info:
  title: VegBankAPI
  version: '0.1'
servers:
  - url: https://api.server.test/v1
paths:
  /vegbank/plot-observations/{accession_code}:
    get:
      summary: returns a specific plot observation via the observation accession code
      x-status: "beta"
      parameters:
        - in: path
          name: accession_code
          x-status: "beta"
          schema: 
            type: string
          required: true
        - in: query
          name: detail
          x-status: "planned"
          schema: 
            type: string
          description: How much detail to return. Options are "full" or "minimal" and full is default.
      responses: 
        "200":
          description: "found"
          content:
            application/json: 
              schema:
                $ref: "#/components/schemas/plot_observation" #Components need updating. For more current model see vegbank loader module. 
  /vegbank/plot-observations:
    get: 
      summary: returns all plot observations in paginated format.
      x-status: "beta"
      parameters: 
        - in: query
          name: limit
          x-status: "beta"
          schema: 
            type: integer
        - in: query
          name: offset
          x-status: "beta"
          schema: 
            type: integer
        - in: query
          name: detail
          x-status: "beta"
          schema: 
            type: string
          description: How much detail to return. Options are "full" or "minimal" and full is default. 
      responses:
        "200":
          description: "found"
          content: 
            application/json:
              schema: 
                type: array
                items:
                  $ref: "#/components/schemas/plot_observation"
    post: 
      summary: takes a list of plot/observation objects and inserts them into the database. 
      x-status: "planned"
      requestBody:
        content:
          multipart/form-data:
            schema: 
              type: array
              items: 
                $ref: "#/components/schemas/plot_observation"

      responses:
        "200": 
          description: "completed"
          content: 
            application/json:
              schema: 
                type: array
                items: 
                  type: string
                  description: accession codes for the new created records
  /vegbank/community-classifications:
    get: 
      summary: returns all community classes in paginated format. 
      x-status: "beta"
      parameters: 
        - in: query
          name: limit
          x-status: "beta"
          schema: 
            type: integer
        - in: query
          name: offset
          x-status: "beta"
          schema: 
            type: integer
        - in: query
          name: detail
          x-status: "planned"
          schema:
            type: string
          description: How much detail to return. Options are "full" or "minimal" and full is default.
      responses:
        "200": 
          description: "found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/community_class" #Components need updating. For more current model see vegbank loader module.
  /vegbank/community-classifications/{accession_code}: 
    get: 
      summary: returns a specific community class via its accession code
      x-status: "beta"
      parameters:
        - in: path
          name: accession_code
          x-status: "beta"
          schema: 
            type: string
          required: true
        - in: query
          name: detail
          x-status: "planned"
          schema: 
            type: string
          description: How much detail to return. Options are "full" or "minimal" and full is default.
      responses:
        "200":
          description: "found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/community_class" #Components need updating. For more current model see vegbank loader module.  
  /vegbank/community-concepts/{accession_code}:
    get: 
      x-status: "beta"
      parameters:
        - in: path
          name: accession_code 
          x-status: "beta"
          schema: 
            type: string
          required: true
        - in: query
          name: detail
          x-status: "planned"
          schema: 
            type: string
          description: How much detail to return. Options are "full" or "minimal" and full is default.
      responses: 
        "200":
          description: "found"
          content:
            application/json: 
              schema:
                $ref: "#/components/schemas/community_concept" #Components need updating. For more current model see vegbank loader module. 
  /vegbank/community-concepts:
    get:
      summary: returns all community concepts in paginated format.  
      x-status: "beta"
      parameters: 
        - in: query
          name: limit
          x-status: "beta"
          schema: 
            type: integer
        - in: query
          name: offset
          x-status: "beta"
          schema: 
            type: integer
        - in: query
          name: detail
          x-status: "planned"
          schema:
            type: string
          description: How much detail to return. Options are "full" or "minimal" and full is default.
      responses: 
        "200": 
          description: "found"
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/community_concept"
    post: 
      summary: takes a list of community concepts and adds them to the database, then returns their new accession codes. 
      x-status: "planned"
      requestBody:
        content:
          multipart/form-data:
            schema: 
              type: array
              items: 
                $ref: "#/components/schemas/community_concept" #Components need updating. For more current model see vegbank loader module. 

      responses:
        "200": 
          description: "completed"
          content: 
            application/json:
              schema: 
                type: array
                items: 
                  type: string
                  description: accession codes for the new created records
  /vegbank/taxon-observations/{accession_code}:
    get:
      summary: returns a specific taxon observation via its (taxon observation) accession code. Currently hitting the view for taxa by cover percentage. 
      x-status: "beta"
      parameters: 
        - in: path
          name: accession_code
          x-status: "beta"
          schema: 
            type: string
          required: true
        - in: query
          name: detail
          x-status: "planned"
          schema: 
            type: string
          description: How much detail to return. Options are "full" or "minimal" and full is default. This param isn't implemented yet. 
      responses: 
        "200":
          description: "found"
          content:
            application/json: 
              schema:
                $ref: "#/components/schemas/taxon" #Components need updating. For more current model see vegbank loader module. 
  /vegbank/taxon-observations:
    get: 
      summary: returns all taxa in paginated format. 
      x-status: "beta"
      parameters: 
        - in: query
          name: limit
          x-status: "beta"
          schema: 
            type: integer
        - in: query
          name: offset
          x-status: "beta"
          schema: 
            type: integer
        - in: query
          name: detail
          x-status: "planned"
          schema: 
            type: string
          description: How much detail to return. Options are "full" or "minimal" and full is default. This param isn't implemented yet.   
        - in: query
          name: numTaxa
          x-status: "beta"
          schema: 
            type: integer
          description: How many topTaxa to return per observation, ordered by max cover. Default is 5.    
      responses: 
        "200": 
          description: "found"
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/taxon" #Components need updating. For more current model see vegbank loader module. 
...
components:
  schemas:
...
    reference: 
      type: object
      properties: 
        short_name: 
          type: string
        title: 
          type: string
        pub_date: 
          type: string
        publisher: 
          type: string
        url: 
          type: string
        accession_code: 
          type: string
...
    community_name:
      type: object
      properties: 
        community_name:
          type: string
        reference: 
          type: object
          $ref: "#/components/schemas/reference"
        date_entered: 
          type: string
          format: date
...
    community_concept:
      type: object
      properties: 
        name: 
          type: string
        reference: 
          type: object
          $ref: "#/components/schemas/reference"
        description: 
          type: string
        community_name: 
          type: object
          $ref: "#/components/schemas/community_name"
        accession_code:
          type: string
...