# vegbank/vegbankapi.py
from flask import Flask, jsonify, request, send_file
import psycopg
from psycopg import connect, ClientCursor
from psycopg.rows import dict_row
import pandas as pd
import pyarrow 
import io
import time
import config

UPLOAD_FOLDER = '/vegbank2/uploads' #For future use with uploading parquet files if necessary
ALLOWED_EXTENSIONS = 'parquet'
QUERIES_FOLDER = 'queries/'

app = Flask(__name__)
app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER

params = config.params

@app.route("/")
def welcome_page():
    return "<h1>Welcome to the VegBank API</h1>"

@app.route("/plot-observations", defaults={'accessioncode': None}, methods=['GET'])
@app.route("/plot-observations/<accessioncode>", methods=['GET'])
def get_plot_observations(accessioncode):
    detail = "full"
    limit = 1000
    offset = 0
    if(request.args.get("detail") != None):
        detail = request.args.get("detail")
    if(request.args.get("limit") != None):
        limit = int(request.args.get("limit"))
    if(request.args.get("offset") != None):
        offset = int(request.args.get("offset"))
    
    countSQL = open(QUERIES_FOLDER + "/plot_observation/get_plot_observations_count.sql", "r").read()


    SQL = ""
    if(accessioncode != None):
        SQL = open(QUERIES_FOLDER + "/plot_observation/get_plot_observation_by_accession_code.sql", "r").read()
        data = (accessioncode, )
    else:
        data = (limit, offset, )
        if(detail == "minimal"):
            SQL = open(QUERIES_FOLDER + "/plot_observation/get_plot_observations_minimal.sql", "r").read()
        else:
            SQL = open(QUERIES_FOLDER + "/plot_observation/get_plot_observations_full.sql", "r").read()
    
    toReturn = {}
    with psycopg.connect(**params, row_factory=dict_row) as conn:
        with conn.cursor() as cur:
            if(accessioncode != None):
                data = (accessioncode, )
            cur.execute(SQL, data)
            toReturn["data"] = cur.fetchall()

            if(accessioncode == None):
                cur.execute(countSQL)
                toReturn["count"] = cur.fetchall()
            else:
                toReturn["count"] = len(toReturn["data"])
        conn.close()   
    return jsonify(toReturn)

@app.route("/taxon-observations", defaults={'accessioncode': None}, methods=['GET', 'POST'])
@app.route("/taxon-observations/<accessioncode>")
def get_taxon_observations(accessioncode):
    detail = "full"
    limit = 100
    offset = 0
    numTaxa = 5
    if(request.args.get("detail") != None):
        detail = request.args.get("detail")
    if(request.args.get("numTaxa") != None):
        numTaxa = int(request.args.get("numTaxa"))
    if(request.args.get("limit") != None):
        limit = int(request.args.get("limit"))
    if(request.args.get("offset") != None):
        offset = int(request.args.get("offset"))
    data = (numTaxa, limit, offset, )
    
    countSQL = open(QUERIES_FOLDER + "/taxon_observation/get_top_taxa_count.sql", "r").read()
    countData = (numTaxa, )

    SQL = ""
    if(accessioncode == None):
        SQL = open(QUERIES_FOLDER + "/taxon_observation/get_top_taxa_coverage.sql", "r").read()  
    else: #TODO This either needs to be an observation accession code, or a taxa one.
        SQL = open(QUERIES_FOLDER + "/taxon_observation/get_taxa_by_accession_code.sql", "r").read()
        data = (accessioncode, )
    toReturn = {}
    with psycopg.connect(**params, row_factory=dict_row) as conn:
        with conn.cursor() as cur:
            cur.execute(SQL, data)
            toReturn["data"] = cur.fetchall()
            print("number of records")

            if(accessioncode == None):
                cur.execute(countSQL, countData)
                toReturn["count"] = cur.fetchall()
            else:
                toReturn["count"] = len(toReturn["data"])
        conn.close()      
    return jsonify(toReturn)

@app.route("/community-classifications", defaults={'accessioncode': None}, methods=['GET', 'POST'])
@app.route("/community-classifications/<accessioncode>")
def get_community_classifications(accessioncode):
    detail = "full"
    limit = 1000
    offset = 0
    if(request.args.get("detail") != None):
        detail = request.args.get("detail")
    if(request.args.get("limit") != None):
        limit = int(request.args.get("limit"))
    if(request.args.get("offset") != None):
        offset = int(request.args.get("offset"))
    
    SQL = ""
    if(accessioncode == None): 
        SQL = open(QUERIES_FOLDER + "/community_classification/get_community_classifications_full.sql", "r").read()
        data = (limit, offset, )
    else:
        SQL = open(QUERIES_FOLDER + "/community_classification/get_community_classification_by_accession_code.sql", "r").read()
        data = (accessioncode, )

    toReturn = {}
    with psycopg.connect(**params, row_factory=dict_row) as conn:
        with conn.cursor() as cur:
            cur.execute(SQL, data)
            toReturn["data"] = cur.fetchall()
        conn.close()    
    return jsonify(toReturn)

@app.route("/community-concepts", defaults={'accessioncode': None}, methods=['GET', 'POST'])
@app.route("/community-concepts/<accessioncode>")
def get_community_concepts(accessioncode):
    detail = "full"
    limit = 1000
    offset = 0
    if(request.args.get("detail") != None):
        detail = request.args.get("detail")
    if(request.args.get("limit") != None):
        limit = int(request.args.get("limit"))
    if(request.args.get("offset") != None):
        offset = int(request.args.get("offset"))

    SQL = ""
    if(accessioncode == None): 
        SQL = open(QUERIES_FOLDER + "/community_concept/get_community_concepts_full.sql", "r").read()
        data = (limit, offset, )
    else:
        SQL = open(QUERIES_FOLDER + "/community_concept/get_community_concept_by_accession_code.sql", "r").read()
        data = (accessioncode, )

    toReturn = {}
    with psycopg.connect(**params, row_factory=dict_row) as conn:
        with conn.cursor() as cur:
            cur.execute(SQL, data)
            toReturn["data"] = cur.fetchall()
        conn.close()    
    return jsonify(toReturn)

@app.route("/parties", defaults={'accessioncode': None}, methods=['GET', 'POST'])
@app.route("/parties/<accessioncode>")
def get_parties(accessioncode):
    detail = "full"
    limit = 1000
    offset = 0
    if(request.args.get("detail") != None):
        detail = request.args.get("detail")
    if(request.args.get("limit") != None):
        limit = int(request.args.get("limit"))
    if(request.args.get("offset") != None):
        offset = int(request.args.get("offset"))

    SQL = ""
    if(accessioncode == None): 
        SQL = open(QUERIES_FOLDER + "/party/get_parties_full.sql", "r").read()
        data = (limit, offset, )
    else:
        SQL = open(QUERIES_FOLDER + "/party/get_party_by_accession_code.sql", "r").read()
        data = (accessioncode, )

    toReturn = {}
    with psycopg.connect(**params, row_factory=dict_row) as conn:
        with conn.cursor() as cur:
            cur.execute(SQL, data)
            toReturn["data"] = cur.fetchall()
        conn.close()    
    return jsonify(toReturn)

#Shiny App Endpoints - These will be retired, leaving them here just until we're done testing. 
@app.route("/get_map_points")
def get_map_points():
    toReturn = {}
    with psycopg.connect(**params, row_factory=dict_row) as conn:
        with conn.cursor() as cur:
            SQL = open(QUERIES_FOLDER + "get_map_points.sql", "r").read() 
            cur.execute(SQL)
            toReturn["data"] = cur.fetchall()
        conn.close()      
    return jsonify(toReturn)

@app.route("/get_observation_table", defaults={'limit': None, 'offset': None}, methods=['GET'])
@app.route("/get_observation_table/<limit>", defaults={'offset': None}, methods=['GET'])
@app.route("/get_observation_table/<limit>/<offset>")
def get_observation_table(limit, offset):
    toReturn = []
    if(limit == None):
        limit = 100
    if(offset == None):
        offset = 0
    limit = int(limit)
    offset = int(offset)
    startTime = time.perf_counter()
    with psycopg.connect(**params, cursor_factory=ClientCursor) as conn:
        with conn.cursor() as cur:
            obsStartTime = time.perf_counter()
            SQL = open(QUERIES_FOLDER + "get_obs_table.sql", "r").read() 
            data = (limit, offset, )
            print(cur.mogrify(SQL, data))
            cur.execute(SQL, data)
            columns = [desc[0] for desc in cur.description]
            for record in cur.fetchall():
                toReturn.append(dict(zip(columns, record)))
            obsEndTime = time.perf_counter()
            print(f"Observations queried and processed in: {obsEndTime - obsStartTime:0.4f} seconds")  
            for record in toReturn:
                taxaStartTime = time.perf_counter()
                SQL = open(QUERIES_FOLDER + "get_top_5_taxa_coverage.sql", "r").read()
                data = (record['plot_id'], )
                cur.execute(SQL, data)
                columns = [desc[0] for desc in cur.description]
                taxa = []
                for taxaRecord in cur.fetchall():
                    taxa.append(dict(zip(columns, taxaRecord)))
                record.update({"taxa": taxa})
                taxaEndTime = time.perf_counter()
                print(f"Taxa queried and processed in: {taxaEndTime - taxaStartTime:0.4f} seconds")  
        conn.close() 
    endTime = time.perf_counter()
    print(f"Observation table queried and processed in: {endTime - startTime:0.4f} seconds")     
    return jsonify(toReturn)

@app.route("/get_observation_details/<accessioncode>")
def get_observation_details(accessioncode):
    toReturn = []
    with psycopg.connect(**params, cursor_factory=ClientCursor) as conn:
        with conn.cursor() as cur:
            SQL = open(QUERIES_FOLDER + "get_observation_details.sql", "r").read() 
            data = (accessioncode, )
            cur.execute(SQL, data)
            columns = [desc[0] for desc in cur.description]
            for record in cur.fetchall():
                toReturn.append(dict(zip(columns, record)))
            if(len(toReturn) != 0):
                taxa = []
                SQL = open(QUERIES_FOLDER + "/taxon_observation/get_taxa_for_observation.sql", "r").read() 
                data = (accessioncode, )
                cur.execute(SQL, data)
                columns = [desc[0] for desc in cur.description]
                for record in cur.fetchall():
                    taxa.append(dict(zip(columns, record)))
                toReturn[0].update({"taxa": taxa})
                communities = []
                SQL = open(QUERIES_FOLDER + "/community_concept/get_community_for_observation.sql", "r").read() 
                data = (accessioncode, )
                cur.execute(SQL, data)
                columns = [desc[0] for desc in cur.description]
                for record in cur.fetchall():
                    communities.append(dict(zip(columns, record)))
                toReturn[0].update({"communities": communities})
            else:
                toReturn.append({"error": "No observation found with that accession code."})
        conn.close()      
    return jsonify(toReturn)

#Utilities
def convert_to_parquet(query, data, conn):
    dataframe = None
    with conn.cursor() as cur:
        SQLString = cur.mogrify(query, data)
        print(SQLString)
        dataframe = pd.read_sql(SQLString, conn)
    df_parquet = dataframe.to_parquet()
    return df_parquet

def allowed_file(filename):
    return '.' in filename and \
           filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS

if __name__ == "__main__":
    app.run(host='0.0.0.0',port=80,debug=True)